move world-calendar logic in lib/dates.js to new 'calendars' component
- use Registry in lib.dates.js to invoke world-calendar logic
  on-demand.
