+++ src/lib/dates.js
@@ -9,7 +9,6 @@
 
 'use strict';
 
-var calendars = require('world-calendars');
 var d3 = require('d3');
 var isNumeric = require('fast-isnumeric');
 
@@ -22,6 +21,9 @@ var ONEDAY = constants.ONEDAY;
 var ONEHOUR = constants.ONEHOUR;
 var ONEMIN = constants.ONEMIN;
 var ONESEC = constants.ONESEC;
+var EPOCHJD = constants.EPOCHJD;
+
+var Registry = require('../registry');
 
 var utcFormat = d3.time.format.utc;
 
@@ -30,84 +32,13 @@ var DATETIME_REGEXP = /^\s*(-?\d\d\d\d|\d\d)(-(\d?\d)(-(\d?\d)([ Tt]([01]?\d|2[0
 // for 2-digit years, the first year we map them onto
 var YFIRST = new Date().getFullYear() - 70;
 
-// cache world calendars, so we don't have to reinstantiate
-// during each date-time conversion
-var allCals = {};
-function getCal(calendar) {
-    var calendarObj = allCals[calendar];
-    if(calendarObj) return calendarObj;
-
-    calendarObj = allCals[calendar] = calendars.instance(calendar);
-    return calendarObj;
-}
-
 function isWorldCalendar(calendar) {
-    return calendar && typeof calendar === 'string' && calendar !== 'gregorian';
+    return (
+        calendar &&
+        Registry.componentsRegistry.calendars &&
+        typeof calendar === 'string' && calendar !== 'gregorian'
+    );
 }
-
-// For fast conversion btwn world calendars and epoch ms, the Julian Day Number
-// of the unix epoch. From calendars.instance().newDate(1970, 1, 1).toJD()
-var EPOCHJD = 2440587.5;
-
-// each calendar needs its own default canonical tick. I would love to use
-// 2000-01-01 (or even 0000-01-01) for them all but they don't necessarily
-// all support either of those dates. Instead I'll use the most significant
-// number they *do* support, biased toward the present day.
-var CANONICAL_TICK = {
-    gregorian: '2000-01-01',
-    coptic: '2000-01-01',
-    discworld: '2000-01-01',
-    ethiopian: '2000-01-01',
-    hebrew: '5000-01-01',
-    islamic: '1000-01-01',
-    julian: '2000-01-01',
-    mayan: '5000-01-01',
-    nanakshahi: '1000-01-01',
-    nepali: '2000-01-01',
-    persian: '1000-01-01',
-    jalali: '1000-01-01',
-    taiwan: '1000-01-01',
-    thai: '2000-01-01',
-    ummalqura: '1400-01-01'
-};
-// Start on a Sunday - for week ticks
-// Discworld and Mayan calendars don't have 7-day weeks anyway so don't change them.
-// If anyone really cares we can customize the auto tick spacings for these calendars.
-var CANONICAL_SUNDAY = {
-    gregorian: '2000-01-02',
-    coptic: '2000-01-03',
-    discworld: '2000-01-01',
-    ethiopian: '2000-01-05',
-    hebrew: '5000-01-01',
-    islamic: '1000-01-02',
-    julian: '2000-01-03',
-    mayan: '5000-01-01',
-    nanakshahi: '1000-01-05',
-    nepali: '2000-01-05',
-    persian: '1000-01-01',
-    jalali: '1000-01-01',
-    taiwan: '1000-01-04',
-    thai: '2000-01-04',
-    ummalqura: '1400-01-06'
-};
-
-var DFLTRANGE = {
-    gregorian: ['2000-01-01', '2001-01-01'],
-    coptic: ['1700-01-01', '1701-01-01'],
-    discworld: ['1800-01-01', '1801-01-01'],
-    ethiopian: ['2000-01-01', '2001-01-01'],
-    hebrew: ['5700-01-01', '5701-01-01'],
-    islamic: ['1400-01-01', '1401-01-01'],
-    julian: ['2000-01-01', '2001-01-01'],
-    mayan: ['5200-01-01', '5201-01-01'],
-    nanakshahi: ['0500-01-01', '0501-01-01'],
-    nepali: ['2000-01-01', '2001-01-01'],
-    persian: ['1400-01-01', '1401-01-01'],
-    jalali: ['1400-01-01', '1401-01-01'],
-    taiwan: ['0100-01-01', '0101-01-01'],
-    thai: ['2500-01-01', '2501-01-01'],
-    ummalqura: ['1400-01-01', '1401-01-01']
-};
 
 /*
  * dateTick0: get the canonical tick for this calendar
@@ -115,17 +46,26 @@ var DFLTRANGE = {
  * bool sunday is for week ticks, shift it to a Sunday.
  */
 exports.dateTick0 = function(calendar, sunday) {
-    calendar = (isWorldCalendar(calendar) && calendar) || 'gregorian';
-    if(sunday) return CANONICAL_SUNDAY[calendar];
-    return CANONICAL_TICK[calendar];
+    if(isWorldCalendar(calendar)) {
+        return sunday ?
+            Registry.getComponentMethod('calendars', 'CANONICAL_SUNDAY')[calendar] :
+            Registry.getComponentMethod('calendars', 'CANONICAL_TICK')[calendar];
+    }
+    else {
+        return sunday ? '2000-01-02' : '2000-01-01';
+    }
 };
 
 /*
  * dfltRange: for each calendar, give a valid default range
  */
 exports.dfltRange = function(calendar) {
-    calendar = (isWorldCalendar(calendar) && calendar) || 'gregorian';
-    return DFLTRANGE[calendar];
+    if(isWorldCalendar(calendar)) {
+        return Registry.getComponentMethod('calendars', 'DFLTRANGE')[calendar];
+    }
+    else {
+        return ['2000-01-01', '2001-01-01'];
+    }
 };
 
 // is an object a javascript date?
@@ -229,7 +169,10 @@ exports.dateTime2ms = function(s, calendar) {
         if(y.length === 2) return BADNUM;
 
         var cDate;
-        try { cDate = getCal(calendar).newDate(Number(y), m, d); }
+        try {
+            cDate = Registry.getComponentMethod('calendars', 'getCal')(calendar)
+                .newDate(Number(y), m, d);
+        }
         catch(e) { return BADNUM; } // Invalid ... date
 
         if(!cDate) return BADNUM;
@@ -296,7 +239,8 @@ exports.ms2DateTime = function(ms, r, calendar) {
         var dateJD = Math.floor(msRounded / ONEDAY) + EPOCHJD,
             timeMs = Math.floor(mod(ms, ONEDAY));
         try {
-            dateStr = getCal(calendar).fromJD(dateJD).formatDate('yyyy-mm-dd');
+            dateStr = Registry.getComponentMethod('calendars', 'getCal')(calendar)
+                .fromJD(dateJD).formatDate('yyyy-mm-dd');
         }
         catch(e) {
             // invalid date in this calendar - fall back to Gyyyy-mm-dd
@@ -406,66 +350,6 @@ exports.cleanDate = function(v, dflt, calendar) {
  */
 
 /*
- * convert d3 templates to world-calendars templates, so our users only need
- * to know d3's specifiers. Map space padding to no padding, and unknown fields
- * to an ugly placeholder
- */
-var UNKNOWN = '##';
-var d3ToWorldCalendars = {
-    'd': {'0': 'dd', '-': 'd'}, // 2-digit or unpadded day of month
-    'a': {'0': 'D', '-': 'D'}, // short weekday name
-    'A': {'0': 'DD', '-': 'DD'}, // full weekday name
-    'j': {'0': 'oo', '-': 'o'}, // 3-digit or unpadded day of the year
-    'W': {'0': 'ww', '-': 'w'}, // 2-digit or unpadded week of the year (Monday first)
-    'm': {'0': 'mm', '-': 'm'}, // 2-digit or unpadded month number
-    'b': {'0': 'M', '-': 'M'}, // short month name
-    'B': {'0': 'MM', '-': 'MM'}, // full month name
-    'y': {'0': 'yy', '-': 'yy'}, // 2-digit year (map unpadded to zero-padded)
-    'Y': {'0': 'yyyy', '-': 'yyyy'}, // 4-digit year (map unpadded to zero-padded)
-    'U': UNKNOWN, // Sunday-first week of the year
-    'w': UNKNOWN, // day of the week [0(sunday),6]
-    // combined format, we replace the date part with the world-calendar version
-    // and the %X stays there for d3 to handle with time parts
-    '%c': {'0': 'D M m %X yyyy', '-': 'D M m %X yyyy'},
-    '%x': {'0': 'mm/dd/yyyy', '-': 'mm/dd/yyyy'}
-};
-
-function worldCalFmt(fmt, x, calendar) {
-    var dateJD = Math.floor(x + 0.05 / ONEDAY) + EPOCHJD,
-        cDate = getCal(calendar).fromJD(dateJD),
-        i = 0,
-        modifier, directive, directiveLen, directiveObj, replacementPart;
-    while((i = fmt.indexOf('%', i)) !== -1) {
-        modifier = fmt.charAt(i + 1);
-        if(modifier === '0' || modifier === '-' || modifier === '_') {
-            directiveLen = 3;
-            directive = fmt.charAt(i + 1);
-            if(modifier === '_') modifier = '-';
-        }
-        else {
-            directive = modifier;
-            modifier = '0';
-            directiveLen = 2;
-        }
-        directiveObj = d3ToWorldCalendars[directive];
-        if(!directiveObj) {
-            i += directiveLen;
-        }
-        else {
-            // code is recognized as a date part but world-calendars doesn't support it
-            if(directiveObj === UNKNOWN) replacementPart = UNKNOWN;
-
-            // format the cDate according to the translated directive
-            else replacementPart = cDate.formatDate(directiveObj[modifier]);
-
-            fmt = fmt.substr(0, i) + replacementPart + fmt.substr(i + directiveLen);
-            i += replacementPart.length;
-        }
-    }
-    return fmt;
-}
-
-/*
  * modDateFormat: Support world calendars, and add one item to
  * d3's vocabulary:
  * %{n}f where n is the max number of digits of fractional seconds
@@ -482,7 +366,7 @@ function modDateFormat(fmt, x, calendar) {
     }
     if(isWorldCalendar(calendar)) {
         try {
-            fmt = worldCalFmt(fmt, x, calendar);
+            fmt = Registry.getComponentMethod('calendars', 'worldCalFmt')(fmt, x, calendar);
         }
         catch(e) {
             return 'Invalid';
@@ -546,7 +430,8 @@ exports.formatDate = function(x, fmt, tr, calendar) {
     if(calendar) {
         try {
             var dateJD = Math.floor((x + 0.05) / ONEDAY) + EPOCHJD,
-                cDate = getCal(calendar).fromJD(dateJD);
+                cDate = Registry.getComponentMethod('calendars', 'getCal')(calendar)
+                    .fromJD(dateJD);
 
             if(tr === 'y') dateStr = yearFormatWorld(cDate);
             else if(tr === 'm') dateStr = monthFormatWorld(cDate);
@@ -618,7 +503,7 @@ exports.incrementMonth = function(ms, dMonth, calendar) {
     if(calendar) {
         try {
             var dateJD = Math.round(ms / ONEDAY) + EPOCHJD,
-                calInstance = getCal(calendar),
+                calInstance = Registry.getComponentMethod('calendars', 'getCal')(calendar),
                 cDate = calInstance.fromJD(dateJD);
 
             if(dMonth % 12) calInstance.add(cDate, dMonth, 'm');
@@ -650,7 +535,10 @@ exports.findExactDates = function(data, calendar) {
         d,
         di;
 
-    var calInstance = isWorldCalendar(calendar) && getCal(calendar);
+    var calInstance = (
+        isWorldCalendar(calendar) &&
+        Registry.getComponentMethod('calendars', 'getCal')(calendar)
+    );
 
     for(var i = 0; i < data.length; i++) {
         di = data[i];
